Apply Joris suggestions
